api/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.3.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==5.1.0

# Vector Database
chromadb==0.6.3

# Utilities
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator==2.3.0

# LangChain & LangGraph
langchain==0.3.14
langchain-core==0.3.28
langchain-openai==0.2.14
langgraph==0.2.59
langgraph-checkpoint==2.0.11

# AI/ML
openai==1.59.7
tiktoken==0.8.0
